Quoter view refactoring, Add new component 'QuoterForm'
